refactor(caldav): split server discovery into CalDAVDiscoverer
Discovery (serverUrl + credentials → list calendars) is now its own class,
separate from CalDAVClientDirect (which talks to one already-known calendar
URL). The Browse modal uses CalDAVDiscoverer directly instead of building a
throwaway sync client, and the client's legacy name-match delegates to it.
serverUrl is now genuinely a discovery input, not a confusing peer of
calendarUrl on the sync client. resolveUrl moved to its own module (shared by
both). No behavior change.
