+++ src/caldav/calDAVClientDirect.test.ts
@@ -7,6 +7,8 @@ const mockConfig: CalDAVConnectionConfig = {
     calendarName: 'Tasks',
 };
 
+// Discovery fixtures — the legacy name-match path in connect() runs discovery
+// (delegated to CalDAVDiscoverer); the discoverer is unit-tested in its own file.
 const PRINCIPAL_XML = `<d:multistatus xmlns:d="DAV:">
   <d:response><d:href>/.well-known/caldav</d:href>
     <d:propstat><d:prop>
@@ -81,110 +83,6 @@ describe('CalDAVClientDirect', () => {
 
         it('should throw when creating VTODO without connection', async () => {
             await expect(client.createVTODO('VTODO data', 'uid-123')).rejects.toThrow('Not connected to calendar server');
-        });
-    });
-
-    describe('parseCalendarsFromXML - pure function XML parsing', () => {
-        it('should extract calendar metadata and convert relative URLs to absolute', () => {
-            const response = `<?xml version="1.0" encoding="UTF-8"?>
-<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
-    <d:response>
-        <d:href>/calendars/user/calendar1/</d:href>
-        <d:propstat>
-            <d:prop>
-                <d:displayname><![CDATA[Reminders]]></d:displayname>
-                <d:resourcetype>
-                    <d:collection/>
-                    <c:calendar/>
-                </d:resourcetype>
-                <c:supported-calendar-component-set>
-                    <c:comp name="VTODO"/>
-                </c:supported-calendar-component-set>
-            </d:prop>
-        </d:propstat>
-    </d:response>
-</d:multistatus>`;
-
-            const calendars = CalDAVClientDirect.parseCalendarsFromXML(response, 'https://caldav.example.com');
-
-            expect(calendars).toHaveLength(1);
-            expect(calendars[0]).toEqual({
-                displayName: 'Reminders',
-                supportsVTODO: true,
-                url: 'https://caldav.example.com/calendars/user/calendar1/'
-            });
-        });
-
-        it('should correctly identify calendars that do not support VTODOs', () => {
-            const response = `<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
-    <d:response>
-        <d:href>/calendars/events/</d:href>
-        <d:propstat>
-            <d:prop>
-                <d:displayname>Events</d:displayname>
-                <d:resourcetype><d:collection/><c:calendar/></d:resourcetype>
-                <c:supported-calendar-component-set>
-                    <c:comp name="VEVENT"/>
-                </c:supported-calendar-component-set>
-            </d:prop>
-        </d:propstat>
-    </d:response>
-</d:multistatus>`;
-
-            const calendars = CalDAVClientDirect.parseCalendarsFromXML(response, 'https://caldav.example.com');
-
-            expect(calendars[0].supportsVTODO).toBe(false);
-        });
-
-        it('should filter out non-calendar resources', () => {
-            const response = `<d:multistatus xmlns:d="DAV:">
-    <d:response>
-        <d:href>/principals/user/testuser/</d:href>
-        <d:propstat>
-            <d:prop>
-                <d:resourcetype><d:collection/><d:principal/></d:resourcetype>
-            </d:prop>
-        </d:propstat>
-    </d:response>
-</d:multistatus>`;
-
-            const calendars = CalDAVClientDirect.parseCalendarsFromXML(response, 'https://caldav.example.com');
-            expect(calendars).toHaveLength(0);
-        });
-
-        it('should handle CDATA in displayname', () => {
-            const response = `<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
-    <d:response>
-        <d:href>/cal/</d:href>
-        <d:propstat>
-            <d:prop>
-                <d:displayname><![CDATA[Tasks & Events]]></d:displayname>
-                <d:resourcetype><d:collection/><c:calendar/></d:resourcetype>
-                <c:supported-calendar-component-set><c:comp name="VTODO"/></c:supported-calendar-component-set>
-            </d:prop>
-        </d:propstat>
-    </d:response>
-</d:multistatus>`;
-
-            const calendars = CalDAVClientDirect.parseCalendarsFromXML(response, 'https://caldav.example.com');
-            expect(calendars[0].displayName).toBe('Tasks & Events');
-        });
-
-        it('should use URL as fallback when displayname is missing', () => {
-            const response = `<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
-    <d:response>
-        <d:href>https://caldav.example.com/unnamed-calendar/</d:href>
-        <d:propstat>
-            <d:prop>
-                <d:resourcetype><d:collection/><c:calendar/></d:resourcetype>
-                <c:supported-calendar-component-set><c:comp name="VTODO"/></c:supported-calendar-component-set>
-            </d:prop>
-        </d:propstat>
-    </d:response>
-</d:multistatus>`;
-
-            const calendars = CalDAVClientDirect.parseCalendarsFromXML(response, 'https://caldav.example.com');
-            expect(calendars[0].displayName).toBe('https://caldav.example.com/unnamed-calendar/');
         });
     });
 
@@ -325,111 +223,35 @@ END:VCALENDAR</c:calendar-data>
         });
     });
 
-    describe('parseHrefForProperty - pure function principal/home discovery', () => {
-        it('should extract href when the property element has predeclared namespace', () => {
-            const xml = `<?xml version="1.0"?>
-<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
-  <d:response>
-    <d:href>/principals/user/</d:href>
-    <d:propstat>
-      <d:prop>
-        <c:calendar-home-set><d:href>/calendars/user/</d:href></c:calendar-home-set>
-      </d:prop>
-    </d:propstat>
-  </d:response>
-</d:multistatus>`;
+    describe('connect() and pinned fetch', () => {
+        it('uses calendarUrl directly and makes no discovery requests when pinned', async () => {
+            const request = jest.fn();
+            const pinned = new CalDAVClientDirect(
+                { ...mockConfig, calendarUrl: 'https://caldav.example.com/calendars/user/personal-todos/' },
+                { request },
+            );
 
-            expect(CalDAVClientDirect.parseHrefForProperty(xml, 'calendar-home-set'))
-                .toBe('/calendars/user/');
+            await pinned.connect();
+
+            expect(request).not.toHaveBeenCalled();
+            expect(pinned.isConnected()).toBe(true);
+            expect((pinned as unknown as { calendarUrl: string }).calendarUrl)
+                .toBe('https://caldav.example.com/calendars/user/personal-todos/');
         });
 
-        it('should extract href when SabreDAV/Baïkal declares xmlns inline on the element (issue #71)', () => {
-            const xml = `<?xml version="1.0"?>
-<d:multistatus xmlns:d="DAV:">
-  <d:response>
-    <d:href>/dav.php/principals/user/</d:href>
-    <d:propstat>
-      <d:prop>
-        <cal:calendar-home-set xmlns:cal="urn:ietf:params:xml:ns:caldav">
-          <d:href>/dav.php/calendars/user/</d:href>
-        </cal:calendar-home-set>
-      </d:prop>
-      <d:status>HTTP/1.1 200 OK</d:status>
-    </d:propstat>
-  </d:response>
-</d:multistatus>`;
+        it('connect() without calendarUrl discovers and matches the calendar by name', async () => {
+            const request = jest.fn();
+            mockDiscovery(request);
+            const c = new CalDAVClientDirect({ ...mockConfig, calendarName: 'Personal' }, { request });
 
-            expect(CalDAVClientDirect.parseHrefForProperty(xml, 'calendar-home-set'))
-                .toBe('/dav.php/calendars/user/');
+            await c.connect();
+
+            expect((c as unknown as { calendarUrl: string }).calendarUrl)
+                .toBe('https://caldav.example.com/calendars/user/personal-todos/');
         });
 
-        it('should extract current-user-principal href with inline xmlns', () => {
-            const xml = `<?xml version="1.0"?>
-<d:multistatus xmlns:d="DAV:">
-  <d:response>
-    <d:href>/dav.php/</d:href>
-    <d:propstat>
-      <d:prop>
-        <d:current-user-principal xmlns:d="DAV:">
-          <d:href>/dav.php/principals/user/</d:href>
-        </d:current-user-principal>
-      </d:prop>
-    </d:propstat>
-  </d:response>
-</d:multistatus>`;
-
-            expect(CalDAVClientDirect.parseHrefForProperty(xml, 'current-user-principal'))
-                .toBe('/dav.php/principals/user/');
-        });
-
-        it('should return null when the property is absent', () => {
-            const xml = `<d:multistatus xmlns:d="DAV:"><d:response><d:href>/x/</d:href></d:response></d:multistatus>`;
-            expect(CalDAVClientDirect.parseHrefForProperty(xml, 'calendar-home-set')).toBeNull();
-        });
-    });
-
-  describe('connect(), listCalendars(), and pinned fetch', () => {
-    it('uses calendarUrl directly and makes no discovery requests when pinned', async () => {
-      const request = jest.fn();
-      const pinned = new CalDAVClientDirect(
-        { ...mockConfig, calendarUrl: 'https://caldav.example.com/calendars/user/personal-todos/' },
-        { request },
-      );
-
-      await pinned.connect();
-
-      expect(request).not.toHaveBeenCalled();
-      expect(pinned.isConnected()).toBe(true);
-      expect((pinned as unknown as { calendarUrl: string }).calendarUrl)
-        .toBe('https://caldav.example.com/calendars/user/personal-todos/');
-    });
-
-    it('listCalendars() discovers and returns every calendar with VTODO support flags', async () => {
-      const request = jest.fn();
-      mockDiscovery(request);
-      const c = new CalDAVClientDirect(mockConfig, { request });
-
-      const calendars = await c.listCalendars();
-
-      expect(calendars).toEqual([
-        { url: 'https://caldav.example.com/calendars/user/personal-todos/', displayName: 'Personal', supportsVTODO: true },
-        { url: 'https://caldav.example.com/calendars/user/personal-events/', displayName: 'Personal', supportsVTODO: false },
-      ]);
-    });
-
-    it('connect() without calendarUrl matches the calendar by name', async () => {
-      const request = jest.fn();
-      mockDiscovery(request);
-      const c = new CalDAVClientDirect({ ...mockConfig, calendarName: 'Personal' }, { request });
-
-      await c.connect();
-
-      expect((c as unknown as { calendarUrl: string }).calendarUrl)
-        .toBe('https://caldav.example.com/calendars/user/personal-todos/');
-    });
-
-    it('fetchVTODOs resolves relative hrefs against the pinned URL (empty serverUrl)', async () => {
-      const REPORT_XML = `<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
+        it('fetchVTODOs resolves relative hrefs against the pinned URL (empty serverUrl)', async () => {
+            const REPORT_XML = `<d:multistatus xmlns:d="DAV:" xmlns:c="urn:ietf:params:xml:ns:caldav">
   <d:response>
     <d:href>/calendars/user/personal-todos/t1.ics</d:href>
     <d:propstat><d:prop>
@@ -437,17 +259,17 @@ END:VCALENDAR</c:calendar-data>
     </d:prop></d:propstat>
   </d:response>
 </d:multistatus>`;
-      const request = jest.fn().mockResolvedValueOnce({ status: 207, text: REPORT_XML, headers: {} });
-      const c = new CalDAVClientDirect(
-        { ...mockConfig, serverUrl: '', calendarUrl: 'https://caldav.example.com/calendars/user/personal-todos/' },
-        { request },
-      );
+            const request = jest.fn().mockResolvedValueOnce({ status: 207, text: REPORT_XML, headers: {} });
+            const c = new CalDAVClientDirect(
+                { ...mockConfig, serverUrl: '', calendarUrl: 'https://caldav.example.com/calendars/user/personal-todos/' },
+                { request },
+            );
 
-      await c.connect();
-      const vtodos = await c.fetchVTODOs();
+            await c.connect();
+            const vtodos = await c.fetchVTODOs();
 
-      expect(vtodos).toHaveLength(1);
-      expect(vtodos[0].url).toBe('https://caldav.example.com/calendars/user/personal-todos/t1.ics');
+            expect(vtodos).toHaveLength(1);
+            expect(vtodos[0].url).toBe('https://caldav.example.com/calendars/user/personal-todos/t1.ics');
+        });
     });
-  });
 });
